pyproject.toml
[tool.poetry]
name = "chatwithyoutubegpt"
version = "0.1.0"
description = ""
authors = ["Anatolii Burtsev <ab.canada.world@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aws-cdk-lib = "2.104.0"
constructs = "^10.3.0"
black = "^23.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
